+++ packages/@leight/file-client/src/sdk/ClientSourceProvider.tsx
@@ -55,4 +55,4 @@ export const FileQueryProvider: FC<IFileQueryProviderProps> = props => {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_wtwatrbpn399g9ya8nuyafh4               = true;
+export const $leight_dtp65i0hql4ui6ic6kfsf120               = true;