+++ packages/@leight/file-server/src/sdk/ServerPrismaSource.ts
@@ -72,4 +72,4 @@ export class FileBasePrismaSource extends AbstractSource<IFileSourceSchema> {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_act7hdlilo0770g1mcrw8f0g = true;
+export const $leight_pymmy27x7mc08810wobxvi0v = true;