+++ packages/@leight/file-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const FileSourceProcedure              = withSourceProcedure<IFileSourceS
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_grwglbx70kjxke5f045exvyl = true;
+export const $leight_et9ncvuuoel1qnvzlgmdvi0j = true;