+++ packages/@leight/file/src/sdk/PrismaSchema.ts
@@ -71,4 +71,4 @@ export const FileQuerySchema                  = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_gonmd9y44ilo3g26rtvu3kdr = true;
+export const $leight_kpc56xwgnzyktlqrhc5reyo8 = true;