+++ packages/@leight/file/src/sdk/Source.ts
@@ -19,4 +19,4 @@ export const FileSourceContext                = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_op3qn9wbinctrbpxyx4mlz0j = true;
+export const $leight_erwxumw9areg1g8ehjkcsmn5 = true;