+++ packages/@leight/job-client/src/sdk/ClientStore.ts
@@ -33,4 +33,4 @@ export const JobSortStore                     = createSortContext<IJobSortSchema
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_j5wazdo34gnd1yvbwupcf2ib = true;
+export const $leight_pjuqczzah9ghsjmkkgp12gyo = true;