+++ packages/@leight/job-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const JobSourceProcedure               = withSourceProcedure<IJobSourceSc
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_xfcogvkzjrg6kkyvj0wuzdno = true;
+export const $leight_xwx638sn6jtasuxecem9uh6i = true;