+++ packages/@leight/user-server/src/sdk/ServerPrismaSource.ts
@@ -72,4 +72,4 @@ export class UserBasePrismaSource extends AbstractSource<IUserSourceSchema> {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_j15rk9w539thh2x9u6r084ea = true;
+export const $leight_cxclythnh3ruor5qm0yiztni = true;