+++ packages/@leight/user/src/sdk/PrismaSchema.ts
@@ -71,4 +71,4 @@ export const UserQuerySchema                  = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_zaay96tvayqa4e36yg4g3hsv = true;
+export const $leight_p3g77au6zg6njywhf4m0jzrc = true;