+++ packages/@leight/user/src/sdk/Source.ts
@@ -20,4 +20,4 @@ export const UserSourceContext                = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_s1lzuufjt0a4wj9h4cuq72ft = true;
+export const $leight_z962rsdo32c0m13q2dbpy09f = true;