+++ packages/@leight/context-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/context-client
 
+## 0.5.67
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/context@0.5.22
+    - @leight/utils@0.5.62
+    - @leight/zustand@0.5.59
+
 ## 0.5.66
 
 ### Patch Changes