+++ packages/@leight/filter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/filter
 
+## 0.5.55
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/zod@0.5.5
+
 ## 0.5.54
 
 ### Patch Changes