+++ packages/@leight/job-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @leight/job-client
 
+## 0.5.20
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/job@0.5.69
+    - @leight/sdk@0.5.17
+    - @leight/i18n@0.5.65
+    - @leight/i18n-client@0.5.65
+    - @leight/query-client@0.5.26
+    - @leight/sort-client@0.5.22
+    - @leight/source-client@0.5.29
+    - @leight/table-client@0.5.28
+    - @leight/utils@0.5.62
+
 ## 0.5.19
 
 ### Patch Changes