+++ packages/@leight/job-server/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @leight/job-server
 
+## 0.5.68
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/prisma@0.5.75
+    - @leight/user@0.5.69
+    - @leight/job@0.5.69
+    - @leight/container@0.5.60
+    - @leight/query@0.5.59
+    - @leight/source@0.5.59
+    - @leight/source-server@0.5.59
+    - @leight/trpc-source-server@0.5.13
+    - @leight/utils@0.5.62
+    - @leight/winston@0.5.58
+
 ## 0.5.67
 
 ### Patch Changes