+++ packages/@leight/layout/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/layout
 
+## 0.5.59
+
+### Patch Changes
+
+- Updated packages
+
 ## 0.5.58
 
 ### Patch Changes