+++ packages/@leight/react-cmdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/react-cmdk
 
+## 0.5.59
+
+### Patch Changes
+
+- Updated packages
+
 ## 0.5.58
 
 ### Patch Changes