+++ packages/@leight/trpc-server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leight/trpc-server
 
+## 0.5.64
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/user-server@0.5.72
+    - @leight/user@0.5.69
+    - @leight/container@0.5.60
+    - @leight/trpc@0.5.58
+
 ## 0.5.63
 
 ### Patch Changes