+++ packages/@leight/winston/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/winston
 
+## 0.5.58
+
+### Patch Changes
+
+- Updated packages
+
 ## 0.5.57
 
 ### Patch Changes