+++ packages/@leight/xlsx-import-server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @leight/xlsx-import-server
 
+## 0.5.58
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/file@0.5.66
+    - @leight/job@0.5.69
+    - @leight/container@0.5.60
+    - @leight/import@0.5.58
+    - @leight/trpc-server@0.5.64
+    - @leight/utils@0.5.62
+    - @leight/utils-server@0.5.58
+    - @leight/xlsx-import@0.5.59
+    - @leight/zod@0.5.5
+
 ## 0.5.57
 
 ### Patch Changes