+++ packages/@leight/xlsx-import/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leight/xlsx-import
 
+## 0.5.59
+
+### Patch Changes
+
+- Updated packages
+- Updated dependencies
+    - @leight/job@0.5.69
+    - @leight/import@0.5.58
+    - @leight/trpc@0.5.58
+    - @leight/zod@0.5.5
+
 ## 0.5.58
 
 ### Patch Changes