+++ packages/@leight/zod/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/zod
 
+## 0.5.5
+
+### Patch Changes
+
+- Updated packages
+
 ## 0.5.4
 
 ### Patch Changes